+++ d60.Cirqus.Tests/Contracts/Views/ViewInstanceDeletion.cs
@@ -12,6 +12,7 @@ namespace d60.Cirqus.Tests.Contracts.Views
     [TestFixture(typeof(MsSqlViewManagerFactory), Category = TestCategories.MsSql)]
     [TestFixture(typeof(EntityFrameworkViewManagerFactory), Category = TestCategories.MsSql)]
     [TestFixture(typeof(InMemoryViewManagerFactory))]
+    [TestFixture(typeof(HybridDbViewManagerFactory), Category = TestCategories.MsSql)]
     [Description("Verifies that view managers are capable of deleting a view instance")]
     public class ViewInstanceDeletion<TFactory> : FixtureBase where TFactory : AbstractViewManagerFactory, new()
     {